refactor: convert bit-byte.utils to service class
