+++ src/features/bit-byte/submit-jeopardy.command.ts
@@ -13,8 +13,8 @@ import {
   Privilege,
   PrivilegeCheck,
 } from "../../middleware/privilege.middleware";
+import { BitByteGroupModel } from "../../models/bit-byte.model";
 import type { RoleId } from "../../types/branded.types";
-import { BitByteGroupModel } from "./bit-byte.model";
 
 class SubmitJeopardyCommand extends SlashCommandHandler {
   public override readonly definition = new SlashCommandBuilder()
